Attempted speed improvements, needs second opinion.
I tried to speed up the core computations in the following ways:

1. Added a function ACol, which computes the i'th column of the A
   matrix (but as a list, so calling
		for i to n list ACol(n,i)
	returns the transpose of the A matrix)
2. Memoizing ACol (by placing "memoize" in front of its definition)
3. I realized that the "compositions" builtin function seems to return
   compositions in revlex order.  In other words, the following returns
the set of compositions in lex order:
		for comp in compositions(n,k) list (reverse comp)
	this means we can remove the need to sort compositions by
rewriting the entry calculator to run through the composition in reverse
order.

Implementing these changes resulted in significant speed improvements
when I called faceMapik and degenMapik individually, but not when I
called faceMapi and not when I called the simplicialModule constructor
on a real complex.  It is faster when I call simplicialModule with
degeneracy set to true, because degenMapi is still faster.
  Furthermore, faceMapi becomes slower when ACol is memoized.  I'm very
confused as to why this is the case.  Some thoughts:
1. An obvious guess is that I goofed and that the code for faceMapi is
   somehow still running the same code as before.  I couldn't see any
reason why this would be the case, but I would appreciate another set of
eyes to make sure I didn't miss something stupid.
2. compositions(n-1,j) runs faster after compositions(n,k) has been
   called (for any j,k), so there might be some kind of caching speedup that's
happening in faceMapik that's errasing the speed gain.  I'm not
convinced by this, because calling ABuilder(n,k) doesn't seem to speed up ABuilder(n-1,j) for any values of n,j.
But I can't think of anything else.

I put the modified functions in SimplicialMapUtilities_Modified and
SimplicialModule_Modified, with "Modified" appended at the end of all of
them (faceMapi -> faceMapiModified, faceMapik -> faceMapik -> Modified,
etc). so that testing is easier to do.  Let me know if you have any
questions or if you can spot anything.